Update[Extension]: CGRect
Property Double -> CGFloat

+++ Classes/UIKit/Modal/HXModalPresentAnimation.swift
@@ -30,14 +30,14 @@ class HXModalPresentAnimation: NSObject, UIViewControllerAnimatedTransitioning {
     private(set) var duration: TimeInterval
     public var hasMask: Bool
     public var ratio: Double
-    public var displayHeight: Double
+    public var displayHeight: CGFloat
 
     // MARK: - Private Property -
     private var backgroundView: UIView?
     private var presented: UIViewController?
 
     // MARK: - Init Methods -
-    public init(withDirection direction: HXMoalDirection = .bottom, duration: TimeInterval = HXModalDefaultDuration, hasMask mask: Bool = true, ratio: Double = 1.0, displayHeight: Double = 0) {
+    public init(withDirection direction: HXMoalDirection = .bottom, duration: TimeInterval = HXModalDefaultDuration, hasMask mask: Bool = true, ratio: Double = 1.0, displayHeight: CGFloat = 0) {
         self.direction = direction
         self.duration = duration
         self.hasMask = mask
@@ -83,7 +83,7 @@ class HXModalPresentAnimation: NSObject, UIViewControllerAnimatedTransitioning {
                     toView.startCenter(byModalDirection: direction)
 
                     UIView.animate(withDuration: duration, delay: 0.0, options: .curveEaseInOut, animations: {
-                        let ratio = (self.displayHeight > 0) ? (self.displayHeight / containerView.frame.height) : self.ratio
+                        let ratio = (self.displayHeight > 0) ? Double(self.displayHeight / containerView.frame.height) : self.ratio
                         toView.endCenter(byModalDirection: self.direction, ratio: ratio)
                     }, completion: { (finished) in
                         let isCancelled = transitionContext.transitionWasCancelled
@@ -117,13 +117,13 @@ extension UIView {
 
         switch direction {
         case .top:
-            center.y = CGFloat(frame.height * Double(-0.5))
+            center.y = frame.height * -0.5
         case .bottom:
-            center.y = CGFloat(frame.height * Double(1.5))
+            center.y = frame.height * 1.5
         case .left:
-            center.x = CGFloat(frame.width * Double(-0.5))
+            center.x = frame.width * -0.5
         case .right:
-            center.x = CGFloat(frame.width * Double(1.5))
+            center.x = frame.width * 1.5
         }
     }
 
@@ -131,13 +131,13 @@ extension UIView {
 
         switch direction {
         case .top:
-            center.y = CGFloat(frame.height * (ratio - 0.5))
+            center.y = frame.height * CGFloat(ratio - 0.5)
         case .bottom:
-            center.y = CGFloat(frame.height * (1.5 - ratio))
+            center.y = frame.height * CGFloat(1.5 - ratio)
         case .left:
-            center.x = CGFloat(frame.width * (ratio - 0.5))
+            center.x = frame.width * CGFloat(ratio - 0.5)
         case .right:
-            center.x = CGFloat(frame.width * (1.5 - ratio))
+            center.x = frame.width * CGFloat(1.5 - ratio)
         }
     }
 }
